Тест getDelUserPage, addDelUserPage, chdDelUserPage.
